PLA-22644 - Modified s3 access logging code
