--- remediation_worker/jobs/s3_cloudtrail_no_logging/minimum_policy.json
@@ -1,18 +1,0 @@
-{
-    "Version": "2012-10-17",
-    "Statement": [
-        {
-            "Sid": "CloudtrailS3EnableAccessLogging",
-            "Effect": "Allow",
-            "Action": [
-                "s3:PutBucketLogging",
-                "s3:PutBucketAcl",
-                "s3:GetBucketAcl",
-                "s3:CreateBucket",
-                "s3:ListBucket",
-                "s3:PutEncryptionConfiguration"
-            ],
-            "Resource": "*"
-        }
-    ]
-}
