+++ remediation_worker/jobs/s3_enable_access_logging/README.md
@@ -1,20 +1,22 @@
 # Enable S3 Access Logging
 
-This job enables access logging for an S3 bucket.
+This job enable Access Logging for an S3 Bucket. It checks for a target bucket with name "vss-logging-target-{cloud_account_id}-{region}", if the bucket exists then it gives log delivery permissions to that bucket and then stores the logs into it. If the bucket does not exist then it creates one.
 
 ### Applicable Rule
 
 ##### Rule ID:
-5c8c265e7a550e1fb6560c67
+* 5c8c265e7a550e1fb6560c67
+* 5c8c265c7a550e1fb6560c63
 
 ##### Rule Name:
-S3 access logging is not enabled
+* S3 access logging is not enabled
+* Access logging is not enabled for CloudTrail S3 buckets
 
 ## Getting Started
 
 ### Prerequisites
 
-The provided AWS credential must have access to `s3:PutBucketLogging`.
+The provided AWS credential must have access to `s3:PutBucketLogging`, `s3:CreateBucket`, `s3:GetBucketAcl`, `s3:PutBucketAcl`, `s3:ListBucket` and `s3:PutEncryptionConfiguration`.
 
 You may find the latest example policy file [here](minimum_policy.json)
