+++ remediation_worker/jobs/s3_enable_access_logging/minimum_policy.json
@@ -5,6 +5,7 @@
             "Sid": "S3EnableAccessLogging",
             "Effect": "Allow",
             "Action": [
+                "s3:PutEncryptionConfiguration",
                 "s3:PutBucketLogging",
                 "s3:PutBucketAcl",
                 "s3:GetBucketAcl",
