+++ tox.ini
@@ -13,7 +13,6 @@ envlist =
 	unit-rds-backup-retention-30-days
 	unit-elb-enable-access-logs
 	unit-security-group-close-port-22
-	unit-s3-cloudtrail-no-logging
 	unit-azure-network-security-group-close-port-22
 	unit-azure-network-security-group-close-port-3389
 	unit-azure-vm-close-port-22
@@ -111,12 +110,6 @@ description = Unit test the project
 changedir = test
 commands = pytest --capture=no --basetemp="{envtmpdir}" unit/test_elb_enable_access_logs.py
 deps = -r remediation_worker/jobs/elb_enable_access_logs/requirements-dev.txt
-
-[testenv:unit-s3-cloudtrail-no-logging]
-description = Unit test the project
-changedir = test
-commands = pytest --capture=no --basetemp="{envtmpdir}" unit/test_s3_cloudtrail_no_logging.py
-deps = -r remediation_worker/jobs/s3_cloudtrail_no_logging/requirements-dev.txt
 
 [testenv:unit-azure-network-security-group-close-port-22]
 description = Unit test the project
